Change Node.js/npm installation source and flow
Use `n` to manage a system-wide version of Node/npm instead of installing from the package manager
Self-upgrade npm not to the latest version, but to the latest version supported by the Node installation
Check the version of the `node` bin, not the `nodejs` bin. `nodejs` provides `node`, but not vice-versa.